refactor: change code after mentor comments
