+++ src/kanban/README.md
@@ -3,7 +3,7 @@
 Kanban (Japanese: 看板, meaning signboard or billboard) is a scheduling system for lean manufacturing and just-in-time manufacturing (JIT).
 Taiichi Ohno, an industrial engineer at Toyota, developed kanban to improve manufacturing efficiency.
 
-![Image alt](https://raw.githubusercontent.com/efandor/external-courses/ex17_js_components/src/kanban/src/img/kanban.png "Kanban screenshot")
+![Image alt](https://raw.githubusercontent.com/efandor/external-courses/ex17_js_components/src/kanban/src/images/kanban.png "Kanban screenshot")
 
 ## Installation
 
@@ -48,7 +48,7 @@ import { addTask } from './addTask';
 
 `.svg`,`.png`, `.jpg` files processed by `file-loader`
 ```js
-import logo from '../../img/logo.svg';
+import logo from '../../assets/images/logo.svg';
 ```
 
  `.css` files processed by `style-loader` and `css-loader`
