+++ src/kanban/src/Utils/addTask.js
@@ -4,20 +4,20 @@ import { state } from "../components/State/State";
 import {rerender} from './rerender';
 import Cardcss from "../components/Card/card.module.css";
 
-const modal = new Modal('This task is on board alredy');
+const modal = new Modal('This task is on board already');
 
 export const addTask = () => {
     const task = new Task('');
-    const tasksBlock = document.body.querySelector(`.${Cardcss.blockContainer}`);
+    const tasksBlock = document.body.querySelector(`.${Cardcss.taskContainer}`);
 
     task.element.setAttribute('contenteditable', 'true');
     tasksBlock.appendChild(task.element);
     task.element.focus();
 
     const handleTaskName = () => {
-        if (!task.element.textContent) task.element.remove();
-
-        if (task.element.textContent) {
+        if (!task.element.textContent) {
+            task.element.remove()
+        } else {
             let allIssues = [];
    
             state.forEach((card) => allIssues = allIssues.concat(card.issues));
@@ -32,9 +32,7 @@ export const addTask = () => {
                     task.element.remove();
                     modal.element.remove();
                 }, {once:true});
-            }
-
-            if (!isSameTaskName) {
+            } else {
                 task.element.setAttribute('contenteditable', 'false');
                 state[0].issues.push(task.element.textContent);
                 rerender();
