+++ src/kanban/src/components/Card/Card.js
@@ -9,9 +9,9 @@ export class Card {
                         <span class="${css.itemTitle}">${title}</span>
                         <button class="${css.itemMenuTop}">•••</button>
                     </div>
-                    <div class="${css.blockContainer}"></div>
+                    <div class="${css.taskContainer}"></div>
                     <div class="${css.blockDown}">
-                        <button class="${css.itemMenuDown}" id="0">Add card</button>
+                        <button class="${css.addCardButton}">Add card</button>
                     </div>
             `,
         }).element;
