+++ src/kanban/src/components/Card/card.module.css
@@ -25,7 +25,7 @@
   font-size: 18px;
 }
 
-.block-container {
+.task-container {
   flex: 1 0 auto;
   display: flex;
   flex-direction: column;
@@ -48,7 +48,7 @@
 }
 
 
-.item-menu-down {
+.add-card-button {
    border-radius: 5px;
    font-size: 18px;
    color: #5E6C84;
@@ -58,19 +58,19 @@
 }
 
 .item-menu-top:hover,
-.item-menu-down:hover,
+.add-card-button:hover,
 .item-menu-button:hover {
    background: #ffffff;
 }
 
 .block-down .item-menu-top::before,
-.item-menu-down::before,
+.add-card-button::before,
 .item-menu-button::before {
-   content: url(../../img/plus.svg);
+   content: url(../../assets/images/plus.svg);
    padding-right: 5px;
 }
 
-.item-menu-down:disabled {
+.add-card-button:disabled {
    background: transparent;
    cursor: not-allowed;
 }
